[CL-2225] Move analytics service to core
