+++ front/app/modules/commercial/analytics/admin/components/EmailDeliveriesCard/useEmailDeliveries/query.ts
@@ -3,7 +3,10 @@ import { getDateFilter, getInterval } from '../../../utils/query';
 
 // typings
 import { QueryParameters } from './typings';
-import { Query, QuerySchema } from '../../../services/analyticsFacts';
+import {
+  Query,
+  QuerySchema,
+} from '../../../../../../../services/analyticsFacts';
 
 export const query = ({
   startAtMoment,
