+++ front/app/modules/commercial/analytics/admin/components/ProjectStatusCard/config.test.ts
@@ -1,4 +1,4 @@
-import { Query } from '../../services/analyticsFacts';
+import { Query } from '../../../../../../services/analyticsFacts';
 import moment from 'moment';
 import { ProjectStatusCardLabels, projectStatusConfig } from './config';
 import { StatCardData } from '../StatCard/useStatCard/typings';
