+++ front/app/modules/commercial/analytics/admin/components/RegistrationsCard/useRegistrations/query.ts
@@ -6,7 +6,10 @@ import {
 } from '../../../utils/query';
 
 // typings
-import { Query, QuerySchema } from '../../../services/analyticsFacts';
+import {
+  Query,
+  QuerySchema,
+} from '../../../../../../../services/analyticsFacts';
 import { QueryParameters } from './typings';
 
 export const query = ({
