+++ front/app/modules/commercial/analytics/admin/components/VisitorsTrafficSourcesCard/getXlsxData.ts
@@ -1,7 +1,11 @@
 import request from 'utils/request';
 
 // services
-import { apiEndpoint, QuerySchema, Query } from '../../services/analyticsFacts';
+import {
+  apiEndpoint,
+  QuerySchema,
+  Query,
+} from '../../../../../../services/analyticsFacts';
 
 // i18n
 import messages from './messages';
