+++ front/app/services/analyticsFacts.ts
